pyproject.toml
[project]
name = "fruit-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.2.0",
    "gdown>=5.2.0",
    "imagehash>=4.3.2",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
